navmessage: improve handling of stale data
This tries to improve the handling of stale data (data which has
been received in previous subframes but no longer matches the current
OSNMA tags). The timing parameters (ADKD=4) are no longer copied from
one subframe to the next one, since it is relatively easy to gather
the new data in a single subframe (it is only two words and all the
satellites transmit the same information).

For the CED (ADKD=0 and 12), the old data is copied over to the
next subframe, but marked as stale. As long as a word in this new
subframe is received, the stale mark for that particular satellite
is removed (the consistency check with the IODnav still applies
and can remove old data with a distinct IODnav). If the CED is marked
as stale, it counts as invalid.