add Svn type
The Svn type is used to hold an SVN and ensure that the value is
in the appropriate range (1-36) by checking in all the constructors.

This replaces usize and u8 parameters for SVNs in most functions, except
where the SVN is used as PRND, since 255 is also an acceptable value
for PRND (representing all the Galileo constellation).