navmessage: rewrite authentication and use COP
This rewrites the navigation message management. For each tag, the
navigation message corresponding to the previous subframe is fetched
and checked. If the authentication is valid, the authbits are added
to all the navigation messages for the same satellite that have equal
contents, regardless of TOW. This is done so that newer navigation
messages that haven't changed inherit authbits from checking older
versions of them after they have been copied (otherwise the authbits
would only be inherited upon copy).

Additionally, COP is checked, and if the age of the mesage is larger
than the COP the authentication check is not done and the tag is
discarded.