galmon-osnma: discard INAV words from previous subframe
Since galmon INAV words are not strictly ordered in time (due to
different latencies of different receivers), it can happen that we
receive an INAV word from the previous subframe. This would upset
subframe collection logic, which would starting collecting a new
subframe. To avoid this, we discard such words.