galmon-osnma: add workaround for galmon data word 16 tow
During testing I have observed a bug in the data coming from Galmon.
For most (but not all) satellites in E1B, the word that starts at
tow = 29 mod 30 in the subframe, which is a word 16 comes incorrectly
with the wrong tow = 15 mod 30 in the same subframe. This tow
corresponds to the other word 16 in the subframe. (Word 16 is the
reduced CED).

This adds a workaround that looks at the last tow received mod 30
and if we see an INAV word with tow = 15 mod 30 but the last tow
received mod 30 is somewhat larger than 15, we replace the tow by
tow = 29 mod 30 in the same subframe.