implement MACLT and FLX MACSEQ checking
This implements the MAC Look-up Table defined in v1.1 of the OSNMA
ICD, and extends the MACSEQ check to cover the case when there are
FLX entires in the MAC Look-up Table.

These changes are required since the Signal in Space is now using
MACLT 34 since the end of 2023. This MACLT does not appear in ICD
v1.0 and it contains FLX entries, so checks would fail with the
previous code.