use DSM id to distinguish DSM-KROOT and DSM-PKR
Currently, all the DSMs were parsed as DSM-KROOT. The DSM ID must
be used to distingush between DSM-KROOT and DSM-PKR, and parse the
messages correctly. At the moment, PKR is not implemented, so
DSM-PKR messages are not parsed (but at least they are not wrongly
parsed as DSM-KROOT).

To do this, the DSM collector needs to return the DSM ID as well
as the data, so a new structure that contains both elements has
been added.