tesla: remove chain status from Chain
This removes the chain status from the Chain. Since the chain status
can change from don't use to operational/test at any moment but the
Chain is associated with a key and propagated to the next TESLA key
after validation, it does not make sense to store the status in the
Chain.