navmessage: fix NavMessageData gst
The current implementation always returned the GST for the current
subframe as the gst of a NavMessageData object. This is not very
useful.

This describes the gst method more precisely: it should return the
GST of the subframe in which the most recent word that forms part
of this navigation data was received. With this definition, the gst
indicates a timestamp at which we have the certainty that the message
was still the current version.

The gst is computed by subtracting the minimum age of the words of
the navigation data from the current GST.