filter out INAV dummy messages
These INAV pages don't contain valid OSNMA data, according to the ICD.
Thanks to Cillian O'Driscoll for mentioning that this is needed for
the OSNMA Test Vectors to be processed correctly.

https://mastodon.social/@CillianGNSS/111799429189364771

Alert Pages don't contain valid OSNMA data either, but there is no
way to filter this out in Galmon data, because the page type bit is
not present.