run_test_vectors.sh: extract Merkle tree 3 pubkey from Merkle tree
The Public Keys given for Merkle 3 in the test vectors are wrong
(they don't correspond to the Merkle tree or to the test vectors
where they are used). The pubkey in the Merkle tree XML file is
correct, so we extract the key from there to run the test vectors
that need this key.