potentially fetch key from previous chain for Slow MAC
When validating the MACSEQ to process Slow MAC tags with the current
TESLA key, the MACSEQ needs to be checked against an older key. If
there has been a chain renewal, this older key can belong to the
previous chain. This adds some logic to the key storage to keep track
of when the chain for a key starts being applicable. This is used
to fetch the key from the other chain if the key GST needed for the
Slow MAC MACSEQ verification is before the start of applicability
of the current chain.

The algorithm included here is not completely foolproof, and for
instance doesn't handle cases when the current chain is changing
between 3 different chain IDs, but it is good enough for the typical
case of chain renewal in which there is only a single chain ID
change in a long period of time.