use NMA status of MACK subframe to validate tags
According to the ICD, the NMA status used in the tag validation
should be the NMA status of the subframe in which the MACK containing
the tag was transmitted, not the NMA status of the subframe in which
the TESLA key used to validate the tag was transmitted. This distinction
is important in the subframes when the NMA status has just changed.

The software was using the TESLA key NMA status, since it is what was
at hand in the moment when tag verification happens. To implement what
the ICD says, the MackStorage is extended to store also the NMA status
together with each MACK. This is slightly wasteful, because the NMA
status is supposed to be the same for all the satellites in the same
subframe, but it simplifies the implementation and also avoids problems
having to do with inconsistent data.