osnma: refactor handling of chain in force in key storage
This uses a different approach to keeping track of the key in force
in the key storage. The new approach fixes failures in the OSNMA
test vector Chain Revocation Step 2.

Instead of keeping track of which is the key of the chain in force,
we keep track of what is the CID of the chain in force. This simplifies
keeping track of the start of applicability, as we can simply set
it whenever the CID changes value. It also handles the case in
Chain Revocation Step 2, when the DSM-PKR is transmitting a KROOT
for a chain that will be in force. In that case, the receiver never
gets a KROOT from the first chain, and so it would never set the
start of applicability for the key of the new chain correctly.