refactor how don't use status is handled
We follow strictly what the ICD says: "navigation data shall not be
authenticated with the provided OSNMA information". With this change,
the only thing that is affected by the don't use status is whether
authentication bits are added to the navigation data as a consequence
of a successful tag verification. The receiver carries out all the
other operations regardless of the don't use status.