update OSNMA test vectors
The GSC has updated the OSNMA test vectors to fix the problem with
the public key for Merkle Tree 3 (the key distributed in the previous
test vector didn't belong to this Merkle Tree). The URL of the test
vectors has changed, and the version of the Receiver Guidelines has
been bumped.