osnma-longan-nano: do not inline initialization
The OsnmaInterface::new function takes up a lot of stack space,
particularly when building with codegen-units = 1, in part due
to missing applications of RVO (around 19 KiB with codegen-units = 1).
If this gets inlined into the main() function, then this large
amount of stack space is spent for the rest of the program and we
risk running out of stack space (which indeed happens with codegen-units
= 1). To avoid this problem, all the initialization code is put into
a function that is marked with #[inline(never)] so that the stack
needed for initialization is freed when the initialization is done.

This change allows building a firmware that works with codegen-units =
1.